refactor add-page-utils to extract sorting utilities
