+++ frontend/packages/console-shared/src/utils/index.ts
@@ -21,3 +21,4 @@ export * from './annotations';
 export * from './xterm-addon-fullscreen';
 export * from './yup-validations';
 export * from './keyword-filter';
+export * from './order-extensions';
